requirements.txt
boto3
botocore
flask
requests
paramiko
scp
python-dotenv
requests
aws-requests-auth